created a helper fucntion help fecth data for all endpoints
