--- data/one-leagues.json
@@ -1,103 +1,0 @@
-{
-  "league": {
-    "id": 4,
-    "name": "Euro Championship",
-    "type": "Cup",
-    "logo": "https://media-3.api-sports.io/football/leagues/4.png"
-  },
-  "country": {
-    "name": "World",
-    "code": null,
-    "flag": null
-  },
-  "seasons": [
-    {
-      "year": 2008,
-      "start": "2008-06-07",
-      "end": "2008-06-29",
-      "current": false,
-      "coverage": {
-        "fixtures": {
-          "events": true,
-          "lineups": true,
-          "statistics_fixtures": false,
-          "statistics_players": false
-        },
-        "standings": false,
-        "players": true,
-        "top_scorers": true,
-        "top_assists": true,
-        "top_cards": true,
-        "injuries": false,
-        "predictions": true,
-        "odds": false
-      }
-    },
-    {
-      "year": 2012,
-      "start": "2012-06-08",
-      "end": "2012-07-01",
-      "current": false,
-      "coverage": {
-        "fixtures": {
-          "events": true,
-          "lineups": true,
-          "statistics_fixtures": false,
-          "statistics_players": false
-        },
-        "standings": false,
-        "players": true,
-        "top_scorers": true,
-        "top_assists": true,
-        "top_cards": true,
-        "injuries": false,
-        "predictions": true,
-        "odds": false
-      }
-    },
-    {
-      "year": 2016,
-      "start": "2016-06-10",
-      "end": "2016-07-10",
-      "current": false,
-      "coverage": {
-        "fixtures": {
-          "events": true,
-          "lineups": true,
-          "statistics_fixtures": true,
-          "statistics_players": true
-        },
-        "standings": true,
-        "players": true,
-        "top_scorers": true,
-        "top_assists": true,
-        "top_cards": true,
-        "injuries": false,
-        "predictions": true,
-        "odds": false
-      }
-    },
-    {
-      "year": 2020,
-      "start": "2019-03-21",
-      "end": "2021-07-11",
-      "current": true,
-      "coverage": {
-        "fixtures": {
-          "events": true,
-          "lineups": true,
-          "statistics_fixtures": true,
-          "statistics_players": true
-        },
-        "standings": true,
-        "players": true,
-        "top_scorers": true,
-        "top_assists": true,
-        "top_cards": true,
-        "injuries": false,
-        "predictions": true,
-        "odds": false
-      }
-    }
-  ]
-}
